# std.compress.flate: rework decompression and delete compression