--- a/lib/std/Io/Reader.zig
+++ b/lib/std/Io/Reader.zig
@@ -1874,3 +1874,7 @@ pub fn writableVectorWsa(
     }
     return .{ i, n };
 }
+
+test {
+    _ = Limited;
+}