--- a/lib/std/compress.zig
+++ b/lib/std/compress.zig
@@ -1,8 +1,7 @@
 //! Compression algorithms.
 
+/// gzip and zlib are here.
 pub const flate = @import("compress/flate.zig");
-pub const gzip = @import("compress/gzip.zig");
-pub const zlib = @import("compress/zlib.zig");
 pub const lzma = @import("compress/lzma.zig");
 pub const lzma2 = @import("compress/lzma2.zig");
 pub const xz = @import("compress/xz.zig");
@@ -14,6 +13,4 @@ test {
     _ = lzma2;
     _ = xz;
     _ = zstd;
-    _ = gzip;
-    _ = zlib;
 }