@mrjbq7 2025-07-30
maybe i'm using it wrong, but this fails:

```zig
const std = @import("std");
const flate = std.compress.flate;

pub fn main() anyerror!void {
    const input_file = try std.fs.cwd().openFile("8mb.gz", .{});
    var input_buffer: [4096]u8 = undefined;
    var input_reader = input_file.reader(&input_buffer);

    var buffer: [flate.max_window_len]u8 = undefined;
    var decompress: flate.Decompress = .init(&input_reader.interface, .gzip, &buffer);

    const output_file = try std.fs.openFileAbsolute("/dev/null", .{ .mode = .write_only });
    var output_buffer: [flate.max_window_len]u8 = undefined;
    var output_writer = output_file.writer(&output_buffer);

    for (0..2000) |i| {
        const b = try decompress.reader.takeByte();
        if (i % 100 == 0) std.debug.print("{d}\n", .{b});
    }

    try output_writer.interface.flush();
}
```

fails with this

```
$ zig-stage3 run repro2.zig
thread 394104 panic: reached unreachable code
zig/std/debug.zig:559:14: 0x104b6be07 in assert (repro2)
    if (!ok) unreachable; // assertion failure
             ^
zig/std/Io/Writer.zig:266:11: 0x104c0f073 in drainPreserve (repro2)
    assert(0 == try w.vtable.drain(w, &.{""}, 1));
          ^
zig/std/Io/Writer.zig:324:26: 0x104c0ce2f in writableSliceGreedyPreserve (repro2)
        try drainPreserve(w, preserve_len);
                         ^
zig/std/Io/Writer.zig:340:56: 0x104c0b3e7 in writableSlicePreserve (repro2)
    const big_slice = try w.writableSliceGreedyPreserve(preserve_len, len);
                                                       ^
zig/std/compress/flate/Decompress.zig:372:45: 0x104c0b17b in writeMatch (repro2)
    const dest = try w.writableSlicePreserve(flate.history_len, length);
                                            ^
zig/std/compress/flate/Decompress.zig:323:39: 0x104c00b47 in readInner (repro2)
                        try writeMatch(w, length, distance);
                                      ^
zig/std/compress/flate/Decompress.zig:151:21: 0x104bfdb27 in stream (repro2)
    return readInner(d, w, limit) catch |err| switch (err) {
                    ^
zig/std/Io/Reader.zig:1061:33: 0x104c0d667 in fillUnbuffered (repro2)
        r.end += r.vtable.stream(r, &writer, .limited(end_cap.len)) catch |err| switch (err) {
                                ^
zig/std/Io/Reader.zig:1046:26: 0x104c0121f in fill (repro2)
    return fillUnbuffered(r, n);
                         ^
zig/std/Io/Reader.zig:1111:13: 0x104bfe32f in peekByte (repro2)
    try fill(r, 1);
            ^
zig/std/Io/Reader.zig:1119:32: 0x104bfd6fb in takeByte (repro2)
    const result = try peekByte(r);
                               ^
repro2.zig:17:49: 0x104bfd1d3 in main (repro2)
        const b = try decompress.reader.takeByte();
                                                ^
zig/std/start.zig:627:37: 0x104bfcda3 in main (repro2)
            const result = root.main() catch |err| {
                                    ^
???:?:?: 0x19ac72b97 in ??? (???)
???:?:?: 0x0 in ??? (???)
[1]    17843 abort      zig run repro2.zig
```

If i make the buffer **3x** max_window_len it works, if i make it **2x** max_window_len it doesn't work.