@andrewrk 2025-07-31
> maybe i'm using it wrong, but this fails:

you were not using it wrong. it's fixed now from cf7a28febbbe877003d8d4f9a13ceb94698c1e3e (already in master)