@dotcarmen 2025-08-07
> - delete std.Io.SeekableStream
>   - Instead, use *std.fs.File.Reader, *std.fs.File.Writer, or std.ArrayListUnmanaged concrete types, because the implementations will be fundamentally different based on whether you are operating on files or memory.

am I to understand correctly that this means there will be no generic seeking implementation in the new io api? want to confirm before i consider opening a PR